Delay, Boundary, and Noise Problem Types

Delay problems add a history function and lag data to the common evolution problem contract. Boundary-value problems add residual callbacks evaluated over the candidate solution or at boundary points. Noise problems request generation or replay of a noise-process trajectory.

Delay Differential Equations

SciMLBase.DDEProblem — Type

Defines a delay differential equation (DDE) problem. Documentation Page: https://docs.sciml.ai/DiffEqDocs/stable/types/dde_types/

Mathematical Specification of a DDE Problem

To define a DDE Problem, you simply need to give the function $f$, the initial condition $u_0$ at time point $t_0$, and the history function $h$ which together define a DDE:

\[\begin{align*} \frac{du}{dt} &= f(u,h,p,t) & (t \geq t_0) \\ u(t_0) &= u_0, \\ u(t) &= h(t) & (t < t_0). \end{align*}\]

$f$ should be specified as f(u, h, p, t) (or in-place as f(du, u, h, p, t)), $u_0$ should be an AbstractArray (or number) whose geometry matches the desired geometry of u, and $h$ should be specified as described below. The history function h is accessed for all delayed values. Note that we are not limited to numbers or vectors for $u_0$; one is allowed to provide $u_0$ as arbitrary matrices / higher dimension tensors as well.

Functional Forms of the History Function

The history function h can be called in the following ways:

h(p, t): out-of-place calculation

h(out, p, t): in-place calculation

h(p, t, deriv::Type{Val{i}}): out-of-place calculation of the ith derivative

h(out, p, t, deriv::Type{Val{i}}): in-place calculation of the ith derivative

h(args...; idxs): calculation of h(args...) for indices idxs

Note that a dispatch for the supplied history function of matching form is required for whichever function forms are used in the user derivative function f.

Declaring Lags

Lags are declared separately from their use. One can use any lag by simply using the interpolant of h at that point. However, one should use caution in order to achieve the best accuracy. When lags are declared, the solvers can be more efficient and accurate, and this is thus recommended.

Neutral and Retarded Delay Differential Equations

Note that the history function specification can be used to specify general retarded arguments, i.e. h(p,α(u,t)). Neutral delay differential equations can be specified by using the deriv value in the history interpolation. For example, h(p,t-τ, Val{1}) returns the first derivative of the history values at time t-τ.

Note that algebraic equations can be specified by using a singular mass matrix.

Problem Type

Constructors

DDEProblem(f[, u0], h, tspan[, p]; <keyword arguments>)
DDEProblem{isinplace,specialize}(f[, u0], h, tspan[, p]; <keyword arguments>)

isinplace optionally sets whether the function is inplace or not. This is determined automatically, but not inferred. specialize optionally controls the specialization level. See Specialization Levels for more details. The default is AutoSpecialize.

For more details on the in-place and specialization controls, see the ODEFunction documentation.

Parameters are optional, and if not given, then a NullParameters() singleton will be used which will throw nice errors if you try to index non-existent parameters. Any extra keyword arguments are passed on to the solvers. For example, if you set a callback in the problem, then that callback will be added in every solve call.

For specifying Jacobians and mass matrices, see the SciMLFunctions interface.

Arguments

f: The function in the DDE.

u0: The initial condition. Defaults to the value h(p, first(tspan)) of the history function evaluated at the initial time point.

h: The history function for the DDE before t0.

tspan: The timespan for the problem.

p: The parameters with which function f is called. Defaults to NullParameters.

constant_lags: A collection of constant lags used by the history function h. Defaults to ().

dependent_lags A tuple of functions (u, p, t) -> lag for the state-dependent lags used by the history function h. Defaults to ().

neutral: If the DDE is neutral, i.e., if delays appear in derivative terms.

order_discontinuity_t0: The order of the discontinuity at the initial time point. Defaults to 0 if an initial condition u0 is provided. Otherwise, it is forced to be greater or equal than 1.

kwargs: The keyword arguments passed onto the solves.

Dynamical Delay Differential Equations

Much like the dynamical ODE problem (see the Differential Equation Problem Types page of the SciMLBase interface documentation), a Dynamical DDE is a partitioned DDE of the form:

\[\begin{align*} \frac{dv}{dt} &= f_1(u,t,h) \\ \frac{du}{dt} &= f_2(v,h) \\ \end{align*}\]

Constructors

DynamicalDDEProblem(f1, f2[, v0, u0], h, tspan[, p]; <keyword arguments>)
DynamicalDDEProblem{isinplace}(f1, f2[, v0, u0], h, tspan[, p]; <keyword arguments>)

Parameter isinplace optionally sets whether the function is inplace or not. This is determined automatically, but not inferred.

Arguments

f: The function in the DDE.

v0 and u0: The initial condition. Defaults to the values h(p, first(tspan))... of the history function evaluated at the initial time point.

h: The history function for the DDE before t0. Must return an object with the indices 1 and 2, with the values of v and u respectively.

tspan: The timespan for the problem.

p: The parameters with which function f is called. Defaults to NullParameters.

constant_lags: A collection of constant lags used by the history function h. Defaults to ().

dependent_lags A tuple of functions (v, u, p, t) -> lag for the state-dependent lags used by the history function h. Defaults to ().

neutral: If the DDE is neutral, i.e., if delays appear in derivative terms.

order_discontinuity_t0: The order of the discontinuity at the initial time point. Defaults to 0 if an initial condition u0 is provided. Otherwise, it is forced to be greater or equal than 1.

kwargs: The keyword arguments passed onto the solves.

For dynamical and second order DDEs, the history function will return an object with the indices 1 and 2 defined, where h(p, t_prev)[1] is the value of $f_2(v, u, h, p, t_{\mathrm{prev}})$ and h(p, t_prev)[2] is the value of $f_1(v, u, h, p, t_{\mathrm{prev}})$ (this is for consistency with the ordering of the initial conditions in the constructor). The supplied history function must also return such a 2-index object, which can be accomplished with a tuple (v,u) or vector [v,u].

2nd Order Delay Differential Equations

To define a 2nd Order DDE Problem, you simply need to give the function $f$ and the initial condition $u_0$ which define an DDE:

\[u'' = f(u',u,h,p,t)\]

f should be specified as f(du,u,p,t) (or in-place as f(ddu,du,u,p,t)), and u₀ should be an AbstractArray (or number) whose geometry matches the desired geometry of u. Note that we are not limited to numbers or vectors for u₀; one is allowed to provide u₀ as arbitrary matrices / higher dimension tensors as well.

From this form, a dynamical ODE:

\[\begin{align*} v' &= f(v,u,h,p,t) \\ u' &= v \end{align*}\]

Constructors

SecondOrderDDEProblem(f[, du0, u0], h, tspan[, p]; <keyword arguments>)
SecondOrderDDEProblem{isinplace}(f, [, du0, u0], h, tspan[, p]; <keyword arguments>)

Parameter isinplace optionally sets whether the function is inplace or not. This is determined automatically, but not inferred.

Arguments

f: The function in the DDE.

du0 and u0: The initial condition. Defaults to the values h(p, first(tspan))... of the history function evaluated at the initial time point.

h: The history function for the DDE before t0. Must return an object with the indices 1 and 2, with the values of v and u respectively.

tspan: The timespan for the problem.

p: The parameters with which function f is called. Defaults to NullParameters.

constant_lags: A collection of constant lags used by the history function h. Defaults to ().

dependent_lags A tuple of functions (v, u, p, t) -> lag for the state-dependent lags used by the history function h. Defaults to ().

neutral: If the DDE is neutral, i.e., if delays appear in derivative terms.

order_discontinuity_t0: The order of the discontinuity at the initial time point. Defaults to 0 if an initial condition u0 is provided. Otherwise, it is forced to be greater or equal than 1.

kwargs: The keyword arguments passed onto the solves.

As above, the history function will return an object with indices 1 and 2, with the values of du and u respectively. The supplied history function must also match this return type, e.g. by returning a 2-element tuple or vector.

Example Problems

Example problems can be found in DiffEqProblemLibrary.jl.

To use a sample problem, such as prob_dde_constant_1delay_ip, you can do something like:

#] add DDEProblemLibrary
using DDEProblemLibrary
prob = DDEProblemLibrary.prob_dde_constant_1delay_ip
sol = solve(prob)

SciMLBase.StandardDDEProblem — Type

struct StandardDDEProblem

Marker for standard DDE problem layouts.

StandardDDEProblem() is the default problem_type metadata stored by DDEProblem when the problem is represented directly by a DDE function, history function, initial state, lags, time span, and parameters. Solver code may test problem_type(prob) isa StandardDDEProblem to distinguish this layout from dynamical or second-order DDE encodings.

SciMLBase.AbstractDynamicalDDEProblem — Type

abstract type AbstractDynamicalDDEProblem

Marker supertype for structured DDE problem layouts.

Subtypes identify DDE problems constructed from partitioned first-order dynamics or from second-order dynamics. These markers are available through problem_type on the common DDEProblem representation so solvers can preserve or recover the structured interpretation when needed.

SciMLBase.DynamicalDDEProblem — Type

struct DynamicalDDEProblem{iip} <: SciMLBase.AbstractDynamicalDDEProblem

Marker for partitioned dynamical DDE problem layouts.

DynamicalDDEProblem{iip} is returned by problem_type when a DDE is constructed from two coupled first-order components. The iip parameter records the in-place convention of the underlying DynamicalDDEFunction.

SciMLBase.SecondOrderDDEProblem — Type

struct SecondOrderDDEProblem{iip} <: SciMLBase.AbstractDynamicalDDEProblem

Marker for second-order DDE problem layouts.

SecondOrderDDEProblem{iip} is returned by problem_type when a second-order delay equation is converted to the partitioned DDE form used by DDEProblem. The iip parameter records the in-place convention of the second-derivative function.

SciMLBase.SDDEProblem — Type

Defines a stochastic delay differential equation (SDDE) problem. Documentation Page: https://docs.sciml.ai/DiffEqDocs/stable/types/sdde_types/

Mathematical Specification of a Stochastic Delay Differential Equation (SDDE) Problem

To define a SDDE Problem, you simply need to give the drift function $f$, the diffusion function g, the initial condition $u_0$ at time point $t_0$, and the history function $h$ which together define a SDDE:

\[\begin{align*} du(t) &= f(u,h,p,t) \, dt + g(u,h,p,t) \, dW_t & (t \geq t_0) \\ u(t_0) &= u_0, \\ u(t) &= h(t) & (t < t_0). \end{align*}\]

$f$ should be specified as f(u, h, p, t) (or in-place as f(du, u, h, p, t)) (and $g$ should match). $u_0$ should be an AbstractArray (or number) whose geometry matches the desired geometry of u, and $h$ should be specified as described below. The history function h is accessed for all delayed values. Note that we are not limited to numbers or vectors for $u_0$; one is allowed to provide $u_0$ as arbitrary matrices / higher dimension tensors as well.

Note that this functionality should be considered experimental.

Functional Forms of the History Function

The history function h can be called in the following ways:

h(p, t): out-of-place calculation

h(out, p, t): in-place calculation

h(p, t, deriv::Type{Val{i}}): out-of-place calculation of the ith derivative

h(out, p, t, deriv::Type{Val{i}}): in-place calculation of the ith derivative

h(args...; idxs): calculation of h(args...) for indices idxs

Note that a dispatch for the supplied history function of matching form is required for whichever function forms are used in the user derivative function f.

Declaring Lags

Lags are declared separately from their use. One can use any lag by simply using the interpolant of h at that point. However, one should use caution in order to achieve the best accuracy. When lags are declared, the solvers can more efficiently be more accurate and thus this is recommended.

Neutral, Retarded, and Algebraic Stochastic Delay Differential Equations

Note that the history function specification can be used to specify general retarded arguments, i.e. h(p,α(u,t)). Neutral delay differential equations can be specified by using the deriv value in the history interpolation. For example, h(p,t-τ, Val{1}) returns the first derivative of the history values at time t-τ.

Note that algebraic equations can be specified by using a singular mass matrix.

Problem Type

Constructors

SDDEProblem(f,g[, u0], h, tspan[, p]; <keyword arguments>)
SDDEProblem{isinplace,specialize}(f,g[, u0], h, tspan[, p]; <keyword arguments>)

isinplace optionally sets whether the function is inplace or not. This is determined automatically, but not inferred. specialize optionally controls the specialization level. See Specialization Levels for more details. The default is AutoSpecialize.

For more details on the in-place and specialization controls, see the ODEFunction documentation.

Parameters are optional, and if not given, then a NullParameters() singleton will be used which will throw nice errors if you try to index non-existent parameters. Any extra keyword arguments are passed on to the solvers. For example, if you set a callback in the problem, then that callback will be added in every solve call.

For specifying Jacobians and mass matrices, see the SciMLFunctions interface.

Arguments

f: The drift function in the SDDE.

g: The diffusion function in the SDDE.

u0: The initial condition. Defaults to the value h(p, first(tspan)) of the history function evaluated at the initial time point.

h: The history function for the DDE before t0.

tspan: The timespan for the problem.

p: The parameters with which function f is called. Defaults to NullParameters.

constant_lags: A collection of constant lags used by the history function h. Defaults to ().

dependent_lags A tuple of functions (u, p, t) -> lag for the state-dependent lags used by the history function h. Defaults to ().

neutral: If the DDE is neutral, i.e., if delays appear in derivative terms.

order_discontinuity_t0: The order of the discontinuity at the initial time point. Defaults to 0 if an initial condition u0 is provided. Otherwise, it is forced to be greater or equal than 1.

kwargs: The keyword arguments passed onto the solves.

Boundary-Value Problems

SciMLBase.BVProblem — Type

Defines an BVP problem. Documentation Page: https://docs.sciml.ai/DiffEqDocs/stable/types/bvp_types/

Mathematical Specification of a BVP Problem

To define a BVP Problem, you simply need to give the function $f$ and the initial condition $u_0$ which define an ODE:

\[\frac{du}{dt} = f(u,p,t)\]

along with an implicit function bc which defines the residual equation, where

\[bc(u,p,t) = 0\]

is the manifold on which the solution must live. A common form for this is the two-point BVProblem where the manifold defines the solution at two points:

\[\begin{align*} u(t_0) &= a, \\ u(t_f) &= b \end{align*}\]

Problem Type

Constructors

TwoPointBVProblem{isinplace}(f, bc, u0, tspan, p=NullParameters(); kwargs...)
BVProblem{isinplace}(f, bc, u0, tspan, p=NullParameters(); kwargs...)

or if we have an initial guess function initialGuess(p, t) for the given BVP, we can pass the initial guess to the problem constructors:

TwoPointBVProblem{isinplace}(f, bc, initialGuess, tspan, p=NullParameters(); kwargs...)
BVProblem{isinplace}(f, bc, initialGuess, tspan, p=NullParameters(); kwargs...)

For any BVP problem type, bc must be inplace if f is inplace. Otherwise it must be out-of-place.

If the bvp is a StandardBVProblem (also known as a Multi-Point BV Problem) it must define either of the following functions

bc!(residual, u, p, t)
residual = bc(u, p, t)

where residual computed from the current u. u is an array of solution values where u[i] is at time t[i], while p are the parameters. For a TwoPointBVProblem, t = tspan. For the more general BVProblem, u can be all of the internal time points, and for shooting type methods u=sol the ODE solution. Note that all features of the ODESolution are present in this form. In both cases, the size of the residual matches the size of the initial condition.

If the bvp is a TwoPointBVProblem then bc must be a Tuple (bca, bcb) and each of them must define either of the following sets of functions:

bca!(resid_a, u_a, p)
bcb!(resid_b, u_b, p)

or

resid_a = bca(u_a, p)
resid_b = bcb(u_b, p)

where resid_a and resid_b are the residuals at the two endpoints, u_a and u_b are the solution values at the two endpoints, and p are the parameters.

Parameters are optional, and if not given, then a NullParameters() singleton will be used which will throw nice errors if you try to index non-existent parameters. Any extra keyword arguments are passed on to the solvers. For example, if you set a callback in the problem, then that callback will be added in every solve call.

Fields

f: The function for the ODE.

bc: The boundary condition function.

u0: The initial condition. Either the initial condition for the ODE as an initial value problem, or a Vector of values for $u(t_i)$ for collocation methods.

tspan: The timespan for the problem.

p: The parameters for the problem. Defaults to NullParameters.

lb: The lower bounds for the solution variables. Defaults to nothing.

ub: The upper bounds for the solution variables. Defaults to nothing.

lcons: The lower bounds for the constraint residuals. Defaults to nothing.

ucons: The upper bounds for the constraint residuals. Defaults to nothing.

Construction layout: Query problem_type to obtain either StandardBVProblem or TwoPointBVProblem.

singular_term: The singular term of the problem. Defaults to nothing.

kwargs: The keyword arguments passed onto the solves.

Special Keyword Arguments

nlls: Specify that the BVP is a nonlinear least squares problem. Use Val(true) or Val(false) for type stability. By default this is automatically inferred based on the size of the input and outputs, however this is type unstable for any array type that doesn't store array size as part of type information. If we can't reliably infer this, we set it to Nothing. Downstreams solvers must be setup to deal with this case.

SciMLBase.StandardBVProblem — Type

struct StandardBVProblem

Marker for multi-point first-order BVP layouts.

StandardBVProblem() identifies boundary value problems whose boundary condition is supplied as one residual function over the current solution values and mesh. It is returned by problem_type for a BVProblem that does not use separate endpoint boundary-condition functions.

SciMLBase.TwoPointBVProblem — Type

struct TwoPointBVProblem{iip}

Marker for two-point first-order BVP layouts.

TwoPointBVProblem{iip} identifies BVPs whose boundary condition is supplied as a pair of endpoint residual functions. The iip parameter records the in-place convention of the associated BVPFunction used for type-stable construction.

SciMLBase.SecondOrderBVProblem — Type

Defines a second order BVP problem. Documentation Page: https://docs.sciml.ai/DiffEqDocs/stable/types/bvp_types/

Mathematical Specification of a second order BVP Problem

To define a second order BVP Problem, you simply need to give the function $f$ and the initial condition $u_0$ which define an ODE:

\[u'' = \frac{d^2 u}{dt^2} = f(u', u, p, t)\]

along with an implicit function bc which defines the residual equation, where

\[bc(u', u, p, t) = 0\]

is the manifold on which the solution must live. A common form for this is the two-point SecondOrderBVProblem where the manifold defines the solution at two points:

\[\begin{align*} g(u(t_0), u'(t_0)) &= 0 \\ g(u(t_f), u'(t_f)) &= 0 \end{align*}\]

Problem Type

Constructors

TwoPointSecondOrderBVProblem{isinplace}(f, bc, u0, tspan, p=NullParameters(); kwargs...)
SecondOrderBVProblem{isinplace}(f, bc, u0, tspan, p=NullParameters(); kwargs...)

or if we have an initial guess function initialGuess(p, t) for the given BVP, we can pass the initial guess to the problem constructors:

TwoPointSecondOrderBVProblem{isinplace}(f, bc, initialGuess, tspan, p=NullParameters(); kwargs...)
SecondOrderBVProblem{isinplace}(f, bc, initialGuess, tspan, p=NullParameters(); kwargs...)

For any BVP problem type, bc must be inplace if f is inplace. Otherwise it must be out-of-place.

If the bvp is a StandardSecondOrderBVProblem (also known as a Multi-Point BV Problem) it must define either of the following functions

bc!(residual, du, u, p, t)
residual = bc(du, u, p, t)

where residual computed from the current u. u is an array of solution values where u[i] is at time t[i], while p are the parameters. For a TwoPointBVProblem, t = tspan. For the more general BVProblem, u can be all of the internal time points, and for shooting type methods u=sol the ODE solution. Note that all features of the ODESolution are present in this form. In both cases, the size of the residual matches the size of the initial condition.

If the bvp is a TwoPointSecondOrderBVProblem then bc must be a Tuple (bca, bcb) and each of them must define either of the following sets of functions:

bca!(resid_a, du_a, u_a, p)
bcb!(resid_b, du_b, u_b, p)

or

resid_a = bca(du_a, u_a, p)
resid_b = bcb(du_b, u_b, p)

where resid_a and resid_b are the residuals at the two endpoints, u_a and u_b are the solution values at the two endpoints, du_a and du_b are the derivative of solution values at the two endpoints, and p are the parameters.

Parameters are optional, and if not given, then a NullParameters() singleton will be used which will throw nice errors if you try to index non-existent parameters. Any extra keyword arguments are passed on to the solvers. For example, if you set a callback in the problem, then that callback will be added in every solve call.

Fields

f: The function for the ODE.

bc: The boundary condition function.

u0: The initial condition. Either the initial condition for the ODE as an initial value problem, or a Vector of values for $u(t_i)$ for collocation methods.

tspan: The timespan for the problem.

p: The parameters for the problem. Defaults to NullParameters.

lb: The lower bounds for the solution variables. Defaults to nothing.

ub: The upper bounds for the solution variables. Defaults to nothing.

lcons: The lower bounds for the constraint residuals. Defaults to nothing.

ucons: The upper bounds for the constraint residuals. Defaults to nothing.

Construction layout: Query problem_type to obtain either StandardSecondOrderBVProblem or TwoPointSecondOrderBVProblem.

kwargs: The keyword arguments passed onto the solves.

SciMLBase.StandardSecondOrderBVProblem — Type

struct StandardSecondOrderBVProblem

Marker for multi-point second-order BVP layouts.

StandardSecondOrderBVProblem() identifies second-order BVPs whose boundary condition is supplied as one residual function over derivative values, state values, parameters, and mesh points. It is returned by problem_type for a SecondOrderBVProblem when endpoint residual functions are not used.

SciMLBase.TwoPointSecondOrderBVProblem — Type

struct TwoPointSecondOrderBVProblem{iip}

Marker for two-point second-order BVP layouts.

TwoPointSecondOrderBVProblem{iip} identifies second-order BVPs whose boundary condition is supplied as a pair of endpoint residual functions. The iip parameter records the in-place convention of the associated DynamicalBVPFunction used for type-stable construction.

Noise Problems

SciMLBase.NoiseProblem — Type

struct NoiseProblem{N<:SciMLBase.AbstractNoiseProcess, T, K} <: SciMLBase.AbstractNoiseProblem

Problem wrapper for sampling or replaying an AbstractNoiseProcess.

NoiseProblem stores the noise process, the time span over which it should be generated, a deterministic seed, and solver keyword arguments. Noise-process solvers use this problem type to construct a saved noise trajectory that follows the common SciML solution interface.

Fields

noise::SciMLBase.AbstractNoiseProcess

tspan::Any

seed::UInt64

kwargs::Any