SciMLSolutions

Definition of the SciMLSolution Interface

All SciMLSolution types are a subset of some AbstractArray. Types with time series (like ODESolution) are subtypes of RecursiveArrayTools.AbstractVectorOfArray and RecursiveArrayTools.AbstractDiffEqArray where appropriate. Types without a time series (like OptimizationSolution) are directly subsets of AbstractArray.

Array Interface

Instead of working on the Vector{uType} directly, we can use the provided array interface.

sol[j]

to access the value at timestep j (if the timeseries was saved), and

sol.t[j]

to access the value of t at timestep j. For multi-dimensional systems, this will address first by component and lastly by time, and thus

sol[i,j]

will be the ith component at timestep j. Hence, sol[j][i] == sol[i, j]. This is done because Julia is column-major, so the leading dimension should be contiguous in memory. If the independent variables had shape (for example, was a matrix), then i is the linear index. We can also access solutions with shape:

sol[i,k,j]

gives the [i,k] component of the system at timestep j. The colon operator is supported, meaning that

sol[i,:]

gives the timeseries for the ith component.

Common Field Names

u: the solution values

t: the independent variable values, matching the length of the solution, if applicable

resid: the residual of the solution, if applicable

original: the solution object from the original solver, if it's a wrapper algorithm

retcode: see the documentation section on return codes

prob: the problem that was solved

alg: the algorithm used to solve the problem

Return Codes (RetCodes)

The solution types have a retcode field which returns a symbol signifying the error state of the solution. The retcodes are as follows:

:Default: The solver did not set retcodes.

:Success: The integration completed without erroring or the steady state solver from SteadyStateDiffEq found the steady state.

:Terminated: The integration is terminated with terminate!(integrator). Note that this may occur by using TerminateSteadyState from the callback library DiffEqCallbacks.

:MaxIters: The integration exited early because it reached its maximum number of iterations.

:DtLessThanMin: The timestep method chose a stepsize which is smaller than the allowed minimum timestep, and exited early.

:Unstable: The solver detected that the solution was unstable and exited early.

:InitialFailure: The DAE solver could not find consistent initial conditions.

:ConvergenceFailure: The internal implicit solvers failed to converge.

:Failure: General uncategorized failures or errors.

Traits

SciMLSolution API

Abstract SciML Solutions

SciMLBase.SciMLSolution — Type

Union of all base solution types.

Uses a Union so that solution types can be <: AbstractArray

SciMLBase.AbstractNoTimeSolution — Type

abstract type AbstractNoTimeSolution{T, N} <: AbstractArray{T, N}

SciMLBase.AbstractTimeseriesSolution — Type

abstract type AbstractTimeseriesSolution{T, N, A} <: RecursiveArrayTools.AbstractDiffEqArray{T, N, A}

SciMLBase.AbstractNoiseProcess — Type

abstract type AbstractNoiseProcess{T, N, A, isinplace} <: RecursiveArrayTools.AbstractDiffEqArray{T, N, A}

SciMLBase.AbstractEnsembleSolution — Type

abstract type AbstractEnsembleSolution{T, N, A} <: RecursiveArrayTools.AbstractVectorOfArray{T, N, A}

SciMLBase.AbstractLinearSolution — Type

abstract type AbstractLinearSolution{T, N} <: SciMLBase.AbstractNoTimeSolution{T, N}

SciMLBase.AbstractNonlinearSolution — Type

abstract type AbstractNonlinearSolution{T, N} <: SciMLBase.AbstractNoTimeSolution{T, N}

Missing docstring.

Missing docstring for SciMLBase.AbstractQuadratureSolution. Check Documenter's build log for details.

SciMLBase.AbstractSteadyStateSolution — Type

abstract type AbstractNonlinearSolution{T, N} <: SciMLBase.AbstractNoTimeSolution{T, N}

SciMLBase.AbstractAnalyticalSolution — Type

abstract type AbstractAnalyticalSolution{T, N, S} <: SciMLBase.AbstractTimeseriesSolution{T, N, S}

SciMLBase.AbstractODESolution — Type

abstract type AbstractODESolution{T, N, S} <: SciMLBase.AbstractTimeseriesSolution{T, N, S}

SciMLBase.AbstractDDESolution — Type

abstract type AbstractDDESolution{T, N, S} <: SciMLBase.AbstractODESolution{T, N, S}

SciMLBase.AbstractRODESolution — Type

abstract type AbstractRODESolution{T, N, S} <: SciMLBase.AbstractODESolution{T, N, S}

SciMLBase.AbstractDAESolution — Type

abstract type AbstractDAESolution{T, N, S} <: SciMLBase.AbstractODESolution{T, N, S}

Concrete SciML Solutions

SciMLBase.LinearSolution — Type

struct LinearSolution{T, N, uType, R, A, C} <: SciMLBase.AbstractLinearSolution{T, N}

Representation of the solution to an linear system Ax=b defined by a LinearProblem

Fields

u: the representation of the optimization's solution.

resid: the residual of the solver, if the method is an iterative method.

alg: the algorithm type used by the solver.

iters: the number of iterations used to solve the equation, if the method is an iterative method.

retcode: the return code from the solver. Used to determine whether the solver solved successfully (sol.retcode === :Success), whether it terminated due to a user-defined callback (sol.retcode === :Terminated), or whether it exited due to an error. For more details, see the return code section of the DifferentialEquations.jl documentation.

cache: the LinearCache object containing the solver's internal cached variables. This is given to allow continuation of solver usage, for example, solving Ax=b with the same A and a new b without refactorizing A. See the caching interface tutorial for details on how to use the cache effectively: http://linearsolve.sciml.ai/dev/tutorials/caching_interface/

Missing docstring.

Missing docstring for SciMLBase.QuadratureSolution. Check Documenter's build log for details.

SciMLBase.DAESolution — Type

struct DAESolution{T, N, uType, duType, uType2, DType, tType, P, A, ID, DE} <: SciMLBase.AbstractDAESolution{T, N, uType}

Representation of the solution to an differential-algebraic equation defined by an DAEProblem.

DESolution Interface

For more information on interacting with DESolution types, check out the Solution Handling page of the DifferentialEquations.jl documentation.

https://diffeq.sciml.ai/stable/basics/solution/

Fields

u: the representation of the DAE solution. Given as an array of solutions, where u[i] corresponds to the solution at time t[i]. It is recommended in most cases one does not access sol.u directly and instead use the array interface described in the Solution Handling page of the DifferentialEquations.jl documentation.

du: the representation fo the derivatives of the DAE solution.

t: the time points corresponding to the saved values of the DAE solution.

prob: the original DAEProblem that was solved.

alg: the algorithm type used by the solver.

destats: statistics of the solver, such as the number of function evaluations required, number of Jacobians computed, and more.

retcode: the return code from the solver. Used to determine whether the solver solved successfully (sol.retcode === :Success), whether it terminated due to a user-defined callback (sol.retcode === :Terminated), or whether it exited due to an error. For more details, see the return code section of the DifferentialEquations.jl documentation.

SciMLBase.NonlinearSolution — Type

struct NonlinearSolution{T, N, uType, R, P, A, O, uType2} <: SciMLBase.AbstractNonlinearSolution{T, N}

Representation of the solution to an nonlinear equation defined by an NonlinearProblem, or the steady state solution to a differential equation defined by a SteadyStateProblem.

Fields

u: the representation of the nonlinear equation's solution.

resid: the residual of the solution.

prob: the original NonlinearProblem/SteadyStateProblem that was solved.

alg: the algorithm type used by the solver.

original: if the solver is wrapped from an alternative solver ecosystem, such as NLsolve.jl, then this is the original return from said solver library.

retcode: the return code from the solver. Used to determine whether the solver solved successfully (sol.retcode === :Success), whether it terminated due to a user-defined callback (sol.retcode === :Terminated), or whether it exited due to an error. For more details, see the return code section of the DifferentialEquations.jl documentation.

left: if the solver is bracketing method, this is the final left bracket value.

right: if the solver is bracketing method, this is the final right bracket value.

SciMLBase.ODESolution — Type

struct ODESolution{T, N, uType, uType2, DType, tType, rateType, P, A, IType, DE} <: SciMLBase.AbstractODESolution{T, N, uType}

Representation of the solution to an ordinary differential equation defined by an ODEProblem.

DESolution Interface

For more information on interacting with DESolution types, check out the Solution Handling page of the DifferentialEquations.jl documentation.

https://diffeq.sciml.ai/stable/basics/solution/

Fields

u: the representation of the ODE solution. Given as an array of solutions, where u[i] corresponds to the solution at time t[i]. It is recommended in most cases one does not access sol.u directly and instead use the array interface described in the Solution Handling page of the DifferentialEquations.jl documentation.

t: the time points corresponding to the saved values of the ODE solution.

prob: the original ODEProblem that was solved.

alg: the algorithm type used by the solver.

destats: statistics of the solver, such as the number of function evaluations required, number of Jacobians computed, and more.

retcode: the return code from the solver. Used to determine whether the solver solved successfully (sol.retcode === :Success), whether it terminated due to a user-defined callback (sol.retcode === :Terminated), or whether it exited due to an error. For more details, see the return code section of the DifferentialEquations.jl documentation.

SciMLBase.OptimizationSolution — Type

struct OptimizationSolution{T, N, uType, P, A, Tf, O} <: SciMLBase.AbstractOptimizationSolution{T, N}

Representation of the solution to an nonlinear optimization defined by an OptimizationProblem

Fields

u: the representation of the optimization's solution.

prob: the original NonlinearProblem/SteadyStateProblem that was solved.

alg: the algorithm type used by the solver.

original: if the solver is wrapped from an alternative solver ecosystem, such as Optim.jl, then this is the original return from said solver library.

retcode: the return code from the solver. Used to determine whether the solver solved successfully (sol.retcode === :Success), whether it terminated due to a user-defined callback (sol.retcode === :Terminated), or whether it exited due to an error. For more details, see the return code section of the DifferentialEquations.jl documentation.

SciMLBase.RODESolution — Type

struct RODESolution{T, N, uType, uType2, DType, tType, randType, P, A, IType, DE} <: SciMLBase.AbstractRODESolution{T, N, uType}

Representation of the solution to an stochastic differential equation defined by an SDEProblem, or of a random ordinary differential equation defined by an RODEProblem.

DESolution Interface

For more information on interacting with DESolution types, check out the Solution Handling page of the DifferentialEquations.jl documentation.

https://diffeq.sciml.ai/stable/basics/solution/

Fields

u: the representation of the SDE or RODE solution. Given as an array of solutions, where u[i] corresponds to the solution at time t[i]. It is recommended in most cases one does not access sol.u directly and instead use the array interface described in the Solution Handling page of the DifferentialEquations.jl documentation.

t: the time points corresponding to the saved values of the ODE solution.

W: the representation of the saved noise process from the solution. See the Noise Processes page of the DifferentialEquations.jl documentation for more details: https://diffeq.sciml.ai/stable/features/noiseprocess/ . Note that this noise is only saved in full if `savenoise=true` in the solver.

prob: the original SDEProblem/RODEProblem that was solved.

alg: the algorithm type used by the solver.

destats: statistics of the solver, such as the number of function evaluations required, number of Jacobians computed, and more.

retcode: the return code from the solver. Used to determine whether the solver solved successfully (sol.retcode === :Success), whether it terminated due to a user-defined callback (sol.retcode === :Terminated), or whether it exited due to an error. For more details, see the return code section of the DifferentialEquations.jl documentation.